requirements.txt
streamlit>=1.31.0
matplotlib>=3.8.0
numpy==1.26.4
pandas==2.2.1
plotly==5.19.0
python-dateutil==2.8.2
pillow==10.2.0
firebase-admin>=6.8.0
pyrebase4>=4.8.0
python-dotenv>=1.0.0
google-auth>=2.39.0
requests>=2.31.0
pycryptodome>=3.20.0
python-jwt>=4.1.0
google-generativeai>=0.3.0
python-dotenv
Pyrebase4